+++ types/graceful-fs/graceful-fs-tests.ts
@@ -16,7 +16,7 @@ gfs2.chmodSync(buf, 1);
 const gracefulified = gfs.gracefulify(fs);
 const _fs: typeof fs = gracefulified;
 gracefulified.lutimes; // $ExpectType typeof lutimes
-promisify(gracefulified.lutimes); // $ExpectType (path: PathLike, atime: string | number | Date, mtime: string | number | Date) => Promise<void>
+promisify(gracefulified.lutimes); // $ExpectType (path: PathLike, atime: TimeLike, mtime: TimeLike) => Promise<void>
 
 const fseGrace = gfs.gracefulify(fse);
 fseGrace.lutimes; // $ExpectType typeof lutimes
@@ -25,4 +25,4 @@ fs.lutimes(buf, str, str, err => {
     err; // $ExpectType ErrnoException | null
 });
 fs.lutimesSync(buf, str, str);
-promisify(fs.lutimes); // $ExpectType (path: PathLike, atime: string | number | Date, mtime: string | number | Date) => Promise<void>
+promisify(fs.lutimes); // $ExpectType (path: PathLike, atime: TimeLike, mtime: TimeLike) => Promise<void>
