+++ types/node/os.d.ts
@@ -208,6 +208,9 @@ declare module 'os' {
         }
     }
 
+    const devNull: string;
+    const EOL: string;
+
     function arch(): string;
     /**
      * Returns a string identifying the kernel version.
@@ -219,7 +222,6 @@ declare module 'os' {
     function version(): string;
     function platform(): NodeJS.Platform;
     function tmpdir(): string;
-    const EOL: string;
     function endianness(): "BE" | "LE";
     /**
      * Gets the priority of a process.
