+++ types/node/util.d.ts
@@ -200,3 +200,65 @@ declare module 'util' {
 declare module 'node:util' {
     export * from 'util';
 }
+
+declare module 'util/types' {
+    import { KeyObject, webcrypto } from 'crypto';
+
+    function isAnyArrayBuffer(object: unknown): object is ArrayBufferLike;
+    function isArgumentsObject(object: unknown): object is IArguments;
+    function isArrayBuffer(object: unknown): object is ArrayBuffer;
+    function isArrayBufferView(object: unknown): object is NodeJS.ArrayBufferView;
+    function isAsyncFunction(object: unknown): boolean;
+    function isBigInt64Array(value: unknown): value is BigInt64Array;
+    function isBigUint64Array(value: unknown): value is BigUint64Array;
+    function isBooleanObject(object: unknown): object is Boolean;
+    function isBoxedPrimitive(object: unknown): object is String | Number | BigInt | Boolean | Symbol;
+    function isDataView(object: unknown): object is DataView;
+    function isDate(object: unknown): object is Date;
+    function isExternal(object: unknown): boolean;
+    function isFloat32Array(object: unknown): object is Float32Array;
+    function isFloat64Array(object: unknown): object is Float64Array;
+    function isGeneratorFunction(object: unknown): object is GeneratorFunction;
+    function isGeneratorObject(object: unknown): object is Generator;
+    function isInt8Array(object: unknown): object is Int8Array;
+    function isInt16Array(object: unknown): object is Int16Array;
+    function isInt32Array(object: unknown): object is Int32Array;
+    function isMap<T>(
+        object: T | {},
+    ): object is T extends ReadonlyMap<any, any>
+        ? unknown extends T
+            ? never
+            : ReadonlyMap<any, any>
+        : Map<unknown, unknown>;
+    function isMapIterator(object: unknown): boolean;
+    function isModuleNamespaceObject(value: unknown): boolean;
+    function isNativeError(object: unknown): object is Error;
+    function isNumberObject(object: unknown): object is Number;
+    function isPromise(object: unknown): object is Promise<unknown>;
+    function isProxy(object: unknown): boolean;
+    function isRegExp(object: unknown): object is RegExp;
+    function isSet<T>(
+        object: T | {},
+    ): object is T extends ReadonlySet<any>
+        ? unknown extends T
+            ? never
+            : ReadonlySet<any>
+        : Set<unknown>;
+    function isSetIterator(object: unknown): boolean;
+    function isSharedArrayBuffer(object: unknown): object is SharedArrayBuffer;
+    function isStringObject(object: unknown): object is String;
+    function isSymbolObject(object: unknown): object is Symbol;
+    function isTypedArray(object: unknown): object is NodeJS.TypedArray;
+    function isUint8Array(object: unknown): object is Uint8Array;
+    function isUint8ClampedArray(object: unknown): object is Uint8ClampedArray;
+    function isUint16Array(object: unknown): object is Uint16Array;
+    function isUint32Array(object: unknown): object is Uint32Array;
+    function isWeakMap(object: unknown): object is WeakMap<object, unknown>;
+    function isWeakSet(object: unknown): object is WeakSet<object>;
+    function isKeyObject(object: unknown): object is KeyObject;
+    function isCryptoKey(object: unknown): object is webcrypto.CryptoKey;
+}
+
+declare module 'node:util/types' {
+    export * from 'util/types';
+}
